requirements.txt
undetected-geckodriver
selenium-wire